*** empty log message ***

+++ magick/widget.c
@@ -7909,7 +7909,7 @@ MagickExport int XMenuWidget(Display *display,XWindows *windows,
 %  The format of the XMonitorWidget method is:
 %
 %      void XMonitorWidget(Display *display,XWindows *windows,const char *task,
-%        const off_t quantum,const off_t span)
+%        const off_t quantum,const size_t span)
 %
 %  A description of each parameter follows:
 %
@@ -7928,7 +7928,7 @@ MagickExport int XMenuWidget(Display *display,XWindows *windows,
 %
 */
 MagickExport void XMonitorWidget(Display *display,XWindows *windows,
-  const char *task,const off_t quantum,const off_t span)
+  const char *task,const off_t quantum,const size_t span)
 {
   unsigned int
     width;
